WaE: enumeral and non-enumeral type in conditional expression

+++ svx/source/tbxctrls/tbunosearchcontrollers.cxx
@@ -103,7 +103,7 @@ void impl_executeSearch( const css::uno::Reference< css::uno::XComponentContext 
     lArgs[3].Name = OUString(SEARCHITEM_TRANSLITERATEFLAGS);
     SvtCTLOptions aCTLOptions;
     sal_Int32 nFlags = 0;
-    nFlags |= (!aMatchCase ? com::sun::star::i18n::TransliterationModules_IGNORE_CASE : 0);
+    nFlags |= (!aMatchCase ? static_cast<int>(com::sun::star::i18n::TransliterationModules_IGNORE_CASE) : 0);
     nFlags |= (aCTLOptions.IsCTLFontEnabled() ? com::sun::star::i18n::TransliterationModulesExtra::ignoreDiacritics_CTL:0 );
     lArgs[3].Value <<= nFlags;
     lArgs[4].Name = OUString(SEARCHITEM_COMMAND);
